fix(orders): B2B-4859 Repair shared additional schema composition
